projeto/requirements-dev.txt
annotated-types==0.7.0
asgiref==3.7.2
attrs==23.1.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.2.5
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.13.1
django-advanced-filters==2.0.0
django-browser-reload==1.11.0
django-crum==0.7.9
django-debug-toolbar==4.2.0
django-filter==23.3
django-full-crud==0.3
django-import-export==4.1.1
django-object-actions==4.2.0
django-qr-code==4.1.0
django-unfold==0.39.0
django-weasyprint==2.3.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
et-xmlfile==1.1.0
fonttools==4.54.1
html5lib==1.1
idna==3.8
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Markdown==3.4.4
MarkupPy==1.14
mypy==1.5.1
mypy-extensions==1.0.0
novadata-utils==0.3.7
odfpy==1.4.1
openpyxl==3.1.5
pillow==10.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydocstyle==6.3.0
pydyf==0.11.0
pyphen==0.16.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.32.3
rpds-py==0.13.2
segno==1.6.1
simplejson==3.19.2
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.4
static3==0.7.0
tablib==3.5.0
tinycss2==1.3.0
tomli==2.0.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.2.2
weasyprint==62.3
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.2.3
